🚚 Make splitText a util function
