+++ src/utils.js
@@ -6,6 +6,7 @@ export { default as parsePage } from './utils/parsePage';
 export { default as prunePseudo } from './utils/prunePseudo';
 export { default as randomActivity } from './utils/randomActivity';
 export { default as selectorMessage } from './utils/selectorMessage';
+export { default as splitText } from './utils/splitText';
 export { default as toDuration } from './utils/toDuration';
 export { default as toTimestamp } from './utils/toTimestamp';
 export { default as toValidName } from './utils/toValidName';
